Split layers.py into layers module
